knock_knock/pyproject.toml
[tool.poetry]
name = "knock-knock"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"
bpython = "^0.22.1"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
